refactor: records APIルートをクリーンアーキテクチャに準拠させる
- PrismaMedicationRecordRepository を作成（stats/trends のロジックも含む）
- ServerDIContainer に medicationRecordRepository を追加
- 全5 records APIルートをPrisma直接からUsecase/Repository経由に変更
