+++ src/app/api/records/route.ts
@@ -1,26 +1,22 @@
 import { prisma } from '@/lib/prisma';
 import { createRecordSchema } from '@/lib/schemas';
 import { success, created, errorResponse } from '@/lib/auth-helpers';
-import { withAuth, verifyResourceOwnership, validateBodySize, flattenRelations , safeParseJson } from '@/lib/api-helpers';
+import { withAuth, verifyResourceOwnership, validateBodySize, safeParseJson } from '@/lib/api-helpers';
 import { checkRateLimit } from '@/lib/security';
-import { QUERY_LIMITS } from '@/lib/constants';
+import { createServerDIContainer } from '@/infrastructure/ServerDIContainer';
+import { GetMedicationHistory, CreateMedicationRecord } from '@/domain/usecases/ManageMedicationRecords';
 
 export const GET = withAuth(async (userId) => {
   const { allowed } = checkRateLimit(`records-get:${userId}`, { maxAttempts: 30, windowMs: 60000 });
   if (!allowed) return errorResponse('リクエストが多すぎます。しばらくしてから再試行してください。', 429);
-  const records = await prisma.medicationRecord.findMany({
-    where: { userId },
-    orderBy: { takenAt: 'desc' },
-    take: QUERY_LIMITS.DEFAULT,
-    include: {
-      member: { select: { name: true } },
-      medication: { select: { name: true } },
-    },
-  });
-  const result = records.map((r) =>
-    flattenRelations(r, { member: 'memberName', medication: 'medicationName' }),
-  );
-  return success(result);
+
+  const container = createServerDIContainer(userId);
+  const usecase = new GetMedicationHistory(container.medicationRecordRepository);
+  const groups = await usecase.execute();
+
+  // フラットな記録リストとして返す（既存APIとの互換性）
+  const records = groups.flatMap((g) => g.records);
+  return success(records);
 });
 
 export async function POST(request: Request) {
@@ -47,17 +43,18 @@ export async function POST(request: Request) {
     const ownershipError = await verifyResourceOwnership(userId, ownershipChecks);
     if (ownershipError) return ownershipError;
 
-    const record = await prisma.medicationRecord.create({
-      data: {
-        userId,
-        memberId: parsed.data.memberId,
-        medicationId: parsed.data.medicationId,
-        scheduleId: parsed.data.scheduleId,
-        notes: parsed.data.notes,
-        dosageAmount: parsed.data.dosageAmount,
-        ...(parsed.data.takenAt ? { takenAt: new Date(parsed.data.takenAt) } : {}),
-      },
+    const container = createServerDIContainer(userId);
+    const usecase = new CreateMedicationRecord(container.medicationRecordRepository);
+
+    await usecase.execute({
+      memberId: parsed.data.memberId,
+      medicationId: parsed.data.medicationId,
+      scheduleId: parsed.data.scheduleId,
+      notes: parsed.data.notes,
+      dosageAmount: parsed.data.dosageAmount,
+      takenAt: parsed.data.takenAt,
     });
-    return created(record);
+
+    return created({ message: '記録しました' });
   })();
 }
